+++ demo-site/src/main/java/app/WebModule.java
@@ -30,13 +30,13 @@ public class WebModule extends Module {
         site().staticContent("/static");
         site().staticContent("/favicon.ico");
         site().staticContent("/robots.txt");
-        site().message().loadProperties("messages/main.properties");
-        site().message().loadProperties("messages/main_en.properties");
-        site().message().loadProperties("messages/main_en_CA.properties");
-        site().message().language(request -> Optional.of("en_US"), "en_US", "en_CA");
+        site().template().message("messages/main.properties");
+        site().template().message("messages/main_en.properties");
+        site().template().message("messages/main_en_CA.properties");
+        site().template().language(request -> Optional.of("en_US"), "en_US", "en_CA");
 
-        site().template("/template/index.html", IndexPage.class);
-        site().template("/template/upload.html", UploadPage.class);
+        site().template().add("/template/index.html", IndexPage.class);
+        site().template().add("/template/upload.html", UploadPage.class);
 
         IndexController index = bind(IndexController.class);
         route().get("/", index::index);