+++ CHANGELOG.md
@@ -4,8 +4,12 @@
 * use smile for cache/queue?
 * think about Batch, and make it interface if design is final
 * mongo support @EnumValue?
+* make config class statefull to make validation order of config easier? for site().message()/template()
 
 ## Change log
+### 4.2.6
+* config: update template and message config API
+
 ### 4.2.5 (3/2/2016)
 * template: refine i18n language support